fix(territory): type manifest upgrade version
Cast the JSON manifest-version parameter explicitly so PostgreSQL can plan the legacy-season pointer swap on Neon.

+++ server/territory_db.ts
@@ -224,7 +224,7 @@ export class TerritoryRepository {
                   summary = jsonb_build_object(
                     'reason', 'manifest_upgrade',
                     'fromManifestVersion', manifest_version,
-                    'toManifestVersion', $3
+                    'toManifestVersion', $3::integer
                   )
             WHERE id = $1`,
           [season.id, now, this.manifest.version],
